Поправил обработчики, объединил модули с эффектами и масштабированием, изменил название переменных
