+++ js/effects.js
@@ -1,28 +1,50 @@
-import { imgUploadPreview } from './scaling.js';
+const Scale = {
+  MIN : 25,
+  MAX : 100,
+  DEFAULT : 100,
+  STEP : 25
+};
+
+const imagePreview = document.querySelector('.img-upload__preview img');
+const imageScale = document.querySelector('.img-upload__scale');
+const scaleControlDecrease = imageScale.querySelector('.scale__control--smaller');
+const scaleControlIncrease = imageScale.querySelector('.scale__control--bigger');
+const scaleControlValue = imageScale.querySelector('.scale__control--value');
 
 const effectsList = document.querySelector('.effects__list');
 const imgUploadEffectLevel = document.querySelector('.img-upload__effect-level');
 
 imgUploadEffectLevel.classList.add('hidden');
 
-const onFilterClick = (evt) => {
-  imgUploadPreview.className = `effects__preview--${evt.target.value}`;
+let control = Scale.DEFAULT;
+scaleControlValue.value = control;
+
+const onUpScaleButtonClick = () => {
+  control = Math.min (control + Scale.STEP, Scale.MAX);
+  scaleControlValue.value = control;
+  imagePreview.style.transform = `scale(${control / 100})`;
 };
 
-effectsList.addEventListener('click', onFilterClick);
+const onDownScaleButtonClick = () => {
+  control = Math.max (control - Scale.STEP, Scale.MIN);
+  scaleControlValue.value = control;
+  imagePreview.style.transform = `scale(${control / 100})`;
+};
 
-// Наложение эффекта на изображение:
+const onFilterClick = (evt) => {
+  imagePreview.className = `effects__preview--${evt.target.value}`;
+};
 
-// По умолчанию должен быть выбран эффект «Оригинал».
-// На изображение может накладываться только один эффект.
-// При смене эффекта, выбором одного из значений среди радиокнопок .effects__radio, добавить картинке внутри .img-upload__preview CSS-класс, соответствующий эффекту. Например, если выбран эффект .effect-chrome, изображению нужно добавить класс effects__preview--chrome.
-// Интенсивность эффекта регулируется перемещением ползунка в слайдере. Слайдер реализуется сторонней библиотекой для реализации слайдеров noUiSlider. Уровень эффекта записывается в поле .effect-level__value. При изменении уровня интенсивности эффекта (предоставляется API слайдера), CSS-стили картинки внутри .img-upload__preview обновляются следующим образом:
-// Для эффекта «Хром» — filter: grayscale(0..1) с шагом 0.1;
-// Для эффекта «Сепия» — filter: sepia(0..1) с шагом 0.1;
-// Для эффекта «Марвин» — filter: invert(0..100%) с шагом 1%;
-// Для эффекта «Фобос» — filter: blur(0..3px) с шагом 0.1px;
-// Для эффекта «Зной» — filter: brightness(1..3) с шагом 0.1;
-// Для эффекта «Оригинал» CSS-стили filter удаляются.
-// При выборе эффекта «Оригинал» слайдер скрывается.
-// При переключении эффектов, уровень насыщенности сбрасывается до начального значения (100%): слайдер, CSS-стиль изображения и значение поля должны обновляться.
-// На расширенном тарифе регулировку интенсивности эффекта можно не делать, достаточно применить к изображению только эффект. Слайдер при этом нужно скрыть, библиотеку noUiSlider подключать к странице не нужно.
+const addEffectsEventListeners = () => {
+  scaleControlIncrease.addEventListener('click', onUpScaleButtonClick);
+  scaleControlDecrease.addEventListener('click', onDownScaleButtonClick);
+  effectsList.addEventListener('click', onFilterClick);
+};
+
+const removeEffectsEventListeners = () => {
+  scaleControlIncrease.removeEventListener('click', onUpScaleButtonClick);
+  scaleControlDecrease.removeEventListener('click', onDownScaleButtonClick);
+  effectsList.removeEventListener('click', onFilterClick);
+};
+
+export {addEffectsEventListeners , removeEffectsEventListeners };
