+++ js/main.js
@@ -3,5 +3,4 @@ import './pictures.js';
 import './utils.js';
 import './modal.js';
 import './validation.js';
-import './scaling.js';
 import './effects.js';
